fix datatype
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@35725 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/nvram/nvram_linux.c
@@ -33,7 +33,7 @@
 #define PATH_DEV_NVRAM "/dev/nvram"
 #define NVRAM_SPACE_MAGIC			0x50534341	/* 'SPAC' */
 
-static int NVRAMSPACE = NVRAM_SPACE;
+static long NVRAMSPACE = NVRAM_SPACE;
 
 /* Globals */
 static int nvram_fd = -1;
